# api/docs/swagger.yaml
definitions:
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.Pong:
    properties:
      pong:
        type: string
    type: object
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
info:
  contact: {}
  description: This is a example of Social Network
  title: Microservices architecture example
  version: "1.0"
paths:
  /v1/github/{id}:
    get:
      consumes:
      - application/json
      description: Api for test github
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pong'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: TEST GITHUB
      tags:
      - github
  /v1/ping/{id}:
    get:
      consumes:
      - application/json
      description: Api for test ping
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pong'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: TEST PING
      tags:
      - test
  /v1/pong/{id}:
    get:
      consumes:
      - application/json
      description: Api for test pong
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pong'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: TEST PONG
      tags:
      - test
security:
- Auth: []
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"